1. Template Options have been expanded. It is now possible to select both a skeleton and a mesh.
If no skeleton is selected, the tool will process the bones from the
template mesh in the normal fashion.

If the template mesh contains additional bones that are not present in
the template skeleton, the tool will add the extra template mesh bones
to the scene root of the skeleton.

2. Preserve Skeleton Structure:

When enabled, this retains the structure of the template skeleton and
all of its NiNodes (bones).

When disabled, the bones will be flattened during skinning and unused
bones will be deleted.

3. Delete Existing Partitions:

When enabled, this will discard all pre-existing dismemberment partition
information on the target mesh.

When disabled, that data will be retained and used whenever possible

4. Copy Havok Node: When enabled, this has three effects.

4.1: Any NiStringExtraData blocks containing 'Havok' in the name will be
copied from the template mesh.
4.2: 'Preserve Skeleton Structure' setting will be ignored. The
structure of the skeleton will be preserved
4.3: 'Delete Existing Rigging' setting will be ignored. The existing
rigging will be deleted.